requirements.txt
backports.zoneinfo==0.2.1
certifi==2021.5.30
charset-normalizer==2.0.4
hvac==0.11.0
idna==3.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
rfc5424-logging-handler==1.4.3
six==1.16.0
tzlocal==3.0
urllib3==1.26.6
